==> 10_clean.ipynb <==
# Imports

import json
from pandas.io.json import json_normalize

import os
import regex as re
import string
import itertools

import numpy as np
from numpy import mean
import pandas as pd
import matplotlib.pyplot as plt
from pprint import pprint

import nltk
from Bio import pairwise2
from Bio.pairwise2 import format_alignment

import jiwer
from jiwer import wer

import pickle

from sklearn.ensemble import RandomForestClassifier
#nltk.download('words')
import warnings
warnings.filterwarnings("ignore")

# pd.set_option('display.max_rows', 500)
# pd.set_option('display.max_columns', 500)
# pd.set_option('display.width', 1000)
# pd.set_option('display.max_colwidth', -1)

# own functions, placed in functions.py.
from functions import *

# Import dataframe with original data

df = pd.read_pickle("./data/df_raw.pkl")
df.shape 
df_original = df.copy()

# Clean reference texts and hypotesis texts
 - initial volume: 546.592
 - after basic cleaning of ref and hyp: 402.992
 - after dropping Switchboard: 298.896

import regex as re

#------------------------------------
#       BASIC CLEANING
#------------------------------------

#lowercase both reference and hypothesis
df['reference.text'] = df['reference.text'].str.lower()
df['hypothesis.text'] = df['hypothesis.text'].str.lower()

#replace t_v_, t_v_s,  i_d_ with tv, tvs, id
df['reference.text'] = df['reference.text'].str.replace('t_v_', 'tv')
df['reference.text'] = df['reference.text'].str.replace('t_v_s', 'tvs')
df['reference.text'] = df['reference.text'].str.replace('i_d_', 'id')

#remove from reference:, ", [, ], {, }
remove_chars_list = [':', '"', '{', '}', '[', ']', '$']
pattern_remove = '|'.join(['({})'.format(re.escape(c)) for c in remove_chars_list])
df['reference.text'] = df['reference.text'].str.replace(pattern_remove, '')

# remove leading, trailing, multiple spaces
df['reference.text'] = df['reference.text'].apply(lambda x: re.sub('\s+', ' ', x).strip())
df['hypothesis.text'] = df['hypothesis.text'].apply(lambda x: re.sub('\s+', ' ', x).strip())

# remove sentences with <5 words
df = df[df['reference.text'].apply(lambda x: x.count(" ") >= 4)]


#------------------------------------
#       DROP SWITCHBOARD
#------------------------------------

df = df[df['corpus'] != 'switchboard_segmented']


#----------------------------------------------
#       DROP TEXTS WITH MORE THAN 1 SPEAKER
#----------------------------------------------
ref_texts_counts = pd.DataFrame(df['reference.text'].value_counts()).reset_index()
ref_texts_counts.columns = ['reference.text', 'counts']
unique_refs_1_speaker = ref_texts_counts[ref_texts_counts['counts']<=16]['reference.text'].unique()
df = df[df['reference.text'].isin(unique_refs_1_speaker)]

#----------------------------------------------
#       KEEP COLUMNS OF INTEREST
#----------------------------------------------

keep_columns = ['identifier', 'speaker_id', 'file', 'corpus', 'configuration', 'machine', 'reference.text', 'hypothesis.text', 'scoring.wer']
df = df[keep_columns]

keep_columns = ['identifier', 'speaker_id', 'file', 'corpus', 'configuration', 'machine', 'reference.text', 'hypothesis.text', 'scoring.wer', 'recomputed_wer']
df = df[keep_columns]

# From where did we loose data after cleaning?

df_before_cleaning = df_original['corpus'].value_counts()
df_after_cleaning = df['corpus'].value_counts()

df_comparison = pd.concat([df_before_cleaning, df_after_cleaning], axis=1, sort=False)
df_comparison.columns = ['before', 'after']
df_comparison['lost'] = ((df_comparison['before'] - df_comparison['after'])/df_comparison['before']).round(2)
df_comparison.sort_values('before', ascending = False)

# Recompute WER after the cleaning
 - the difference in **mean wer** and **mean recoputed wer** is not significant

#recompute wer after cleaning
import jiwer
from jiwer import wer
df['recomputed_wer'] = df.apply(lambda row: wer(row['reference.text'], row['hypothesis.text']), axis=1)

# average wer and recomputed wer per configuration
df.groupby(['configuration']) \
      .agg(count=('scoring.wer', 'size'), mean_wer=('scoring.wer', 'mean'), mean_recomputed_wer=('recomputed_wer', 'mean')) \
      .reset_index().sort_values('mean_wer')

# average wer and recomputed wer per machine
df.groupby(['machine']) \
      .agg(count=('scoring.wer', 'size'), mean_wer=('scoring.wer', 'mean'), mean_recomputed_wer=('recomputed_wer', 'mean')) \
      .reset_index().sort_values('mean_wer')

# Order data by mean wer by configuration (over all corpora)

# get order of machines (given by avg wer recalculated over all corpora)
mean_wer_agg_recalc = df.groupby(['configuration']) \
       .agg(count=('recomputed_wer', 'size'), mean_recomputed_wer=('recomputed_wer', 'mean')) \
       .reset_index().sort_values('mean_recomputed_wer')

order = mean_wer_agg_recalc['configuration'].values

df['configuration'] = pd.Categorical(df['configuration'],categories=order)
df = df.sort_values('configuration')

# Export clean and enriched data frame

df.to_pickle("./data/df_clean_newwer.pkl")
