==> 20_premodelling.ipynb <==
# Imports

import json
from pandas.io.json import json_normalize

import os
import regex as re
import string
import itertools

import numpy as np
from numpy import mean
import pandas as pd
import matplotlib.pyplot as plt
from pprint import pprint

import nltk
from Bio import pairwise2
from Bio.pairwise2 import format_alignment

import jiwer
from jiwer import wer

import pickle

from sklearn.ensemble import RandomForestClassifier
#nltk.download('words')
import warnings
warnings.filterwarnings("ignore")

# pd.set_option('display.max_rows', 500)
# pd.set_option('display.max_columns', 500)
# pd.set_option('display.width', 1000)
# pd.set_option('display.max_colwidth', -1)

# own functions, placed in functions.py.
from functions import *

# Import cleaned data and prepare for alignment

#import data
df = pd.read_pickle("./data/df_clean_newwer.pkl")
df.shape

machines_keep = ['2', '3', '4', '5']
df = make_reduced_df (df, machines_keep)
df = split_hypothesis (df)

# Prepare train, validation and test data

# select 15% of reference texts for validation and 15% for test
all_references = list(df['reference.text'].unique()) # 18.681 reference texts in total 
reference_texts_train, reference_texts_validation, reference_texts_test = refereces_train_validation_test (all_references, 0.15, 0.15)

# #check for overlaps between lists
# for element in reference_texts_train:
#     matching = [s for s in reference_texts_test if element in s]
# matching

#Export train, validation, test lists
with open('./data/reference_texts_validation.pkl','wb') as f:
    pickle.dump(reference_texts_validation,f)
         
with open('.data./reference_texts_test.pkl','wb') as f:
    pickle.dump(reference_texts_test,f)
         
with open('./data/reference_texts_train.pkl','wb') as f:
    pickle.dump(reference_texts_train,f)
    
#Import train, validation, test lists         
# reference_texts_validation = pickle.load(open('./data/reference_texts_validation.pkl', 'rb'))
# reference_texts_test = pickle.load(open('./data/reference_texts_test.pkl', 'rb'))
# reference_texts_train = pickle.load(open('./data/reference_texts_train.pkl', 'rb'))

df_train_validation = df[~df['reference.text'].isin(reference_texts_test)]
df_test  = df[df['reference.text'].isin(reference_texts_test)]

# # checks
# df_train_validation['reference.text'].nunique() #15.879
# df_test['reference.text'].nunique() #2.802

# Alignment
On the **train and validation**, we align to REFERENCE so that we can compute accuracy. On **test** we align to the BEST configuration.

## Train and validation data - aligned to REF

reference_texts_train_validation = reference_texts_train + reference_texts_validation
alignment_with_ref = alignment(df_train_validation, reference_texts_train_validation, with_reference = True)

alignment_with_ref.to_pickle('./data/alignment_with_ref.pkl')
#alignment_with_ref = pd.read_pickle("./data/alignment_with_ref.pkl")

## Test data - aligned to the hypotehsis of the best machine
We align **all the data** and then we separate the **test data**. The alignment of all the data is useful in order to compute in a next step the performance if majority voting. 

reference_texts = df['reference.text'].unique()
alignment_all = alignment(df, reference_texts, with_reference = False)
alignment_test = alignment_all[alignment_all['reference.text'].isin(reference_texts_test)]

alignment_all.to_pickle('./data/alignment_all.pkl')
#alignment_all = pd.read_pickle("./data/alignment_all.pkl")

alignment_test.to_pickle('./data/alignment_test.pkl')
#alignment_test = pd.read_pickle("./data/alignment_test.pkl")

# Voting

configurations = df['configuration'].unique()
voting_results = vote(alignment_all, reference_texts, configurations)
voting_results

#voting_results.to_pickle('./data/voting_results_all.pkl')
#voting_results.to_csv('./data/voting_results_all.csv')
#voting_results = pd.read_pickle("./data/voting_results_all.pkl")

best_asr_df = make_best_asr_df(df, reference_texts)

assessment_voting = pd.merge(best_asr_df, voting_results, on='reference.text')
assessment_voting['voting_class'] = np.where((assessment_voting['voting_wer'] < assessment_voting['min_wer']) , 'better', np.where((assessment_voting['voting_wer'] > assessment_voting['min_wer']) , 'worse', 'same'))
#assessment_voting.to_pickle('./data/assessment_voting.pkl')
#assessment_voting.to_csv('./data/assessment_voting.csv')
#assessment_voting = pd.read_pickle("./data/assessment_voting.pkl")

assessment_voting

stats_train = stats_voting(assessment_voting, reference_texts_train)
stats_validation = stats_voting(assessment_voting, reference_texts_validation)
stats_test = stats_voting(assessment_voting, reference_texts_test)
stats_overall = stats_voting(assessment_voting, reference_texts)

stats_overall

assessment_voting.mean()
#assessment_voting[assessment_voting['reference.text'].isin(reference_texts_train)].mean()

# Run alignment and voting for one example  
Both to ref and to the best. 

### 1. Some examples to use

## Some examples to use
#example = ['the boy had never even noticed that there was a hole in his pouch']
#example = ['those events led to a domino effect that has yet to end']
#example = ['but they are just shutting up']
#example = ['give me precise instructions how to decontaminate the trooper']
example = ['best solution is to find an area that is predominantly sunlight or shade']
#example = ['these figures of wax look creepy']
#example = ['what difference is there between a bottle and a flagon']
#example = ['you know everyone wants to has some kind of amenities']
#example = ['governments assert that armies are needed above all for external defense but that is not true']
#example = ['actually achieving that in terms of whether the you know the power requirements of the uh such a source um you know compromise the our our need for uh you know it it being']
#example = ['governments assert that armies are needed above all for external defense but that is not true','major foster in a letter to judge george m bennett of minneapolis said']
#example = ['eventually the merchant had to hire two more employees']
#example = ['aligning all the teeth may take a year or more']
## Find more examples:
#df_train_validation['reference.text'].unique()

df[df['reference.text'].isin(example)]

### 2. Alignment to REF

alignment_df = alignment(df, example, with_reference = True)
alignment_df

### 3. Alignment to best

alignment_df = alignment(df, example, with_reference = False)
alignment_df

### 4. voting

voting_results = vote(alignment_df, example, ['5_ae', '5_a', '4_iaebglebg', '4_iabglbg', '3_mdagmlg', '2_kl', '3_mdagrmlg', '2_ka'])
voting_results

### 5. build_mutations function

ref = ['this', 'is', 'a', 'part', 'speech']
hyp = ['part', 'of', 'peech']
m, c = build_mutations(ref, hyp, -1)
m

c
