==> 30_modelling.ipynb <==
import re, string, os
import pandas as pd
import numpy as np
import jiwer
from jiwer import wer
import pickle
import regex as re
import nltk
import itertools
from sklearn.ensemble import RandomForestClassifier
#nltk.download('words')
import warnings
warnings.filterwarnings("ignore")

from functions import *

# Importing data

train = pd.read_pickle("./Data/alignData/ent_with_ref.pkl") 
test = pd.read_pickle("./Data/alignment_test.pkl") 
voting = pd.read_pickle("./Data/assessment_voting.pkl") 

reference_texts_validation = pickle.load(open('./Data/reference_texts_validation.pkl', 'rb'))
reference_texts_test = pickle.load(open('./Data/reference_texts_test.pkl', 'rb'))
reference_texts_train = pickle.load(open('./Data/reference_texts_train.pkl', 'rb'))

# Encoding words

le, col = word_encoding(train, test)
train[col] = train[col].apply(lambda x :le.transform(x))

# Training data preparation

train_data, col_name, machines, match = train_data_v1(train, reference_texts_train, reference_texts_validation)

# Upsampling
train_data = upsampling(train_data)
train_data.to_pickle("./Data/train_data_upsampled.pkl")

# machine tracker creation
train_data_ref = train_data.copy()
train_data = machine_tracker(train_data, col_name, machines, match)
train_data.to_pickle("./Data/train_data_final.pkl")

# Random forest

train_data = pd.read_pickle("./Data/train_data_final.pkl")

X_train, y_train = train_test_split(train_data)
regr = RandomForestClassifier(n_estimators = 20, min_samples_leaf=3, n_jobs=-1)
regr.fit(X_train, y_train)

#pickle.dump(regr, open('random_forest_classifier.sav', 'wb'), protocol=4)

# Test data

test_reduced, col_test, reference_texts_test_reduced = test_data_preparation_1(test, reference_texts_test)
test_reduced[col_test] = test_reduced[col_test].apply(lambda x :le.transform(x))
test_reduced_df, machines_test, match_test = test_data_preparation_2(test_reduced)
test_reduced_df = first_prediction(test_reduced_df, regr)

test_reduced_df2 = test_predictions(test_reduced_df, machines_test, match_test, regr, col_test)
#test_reduced_df2.to_pickle("./Test_predictions_final(full)_pickl.pkl")

# Comparison Analysis

ori_df = pd.read_pickle("./Data/Dataframe/df_clean.pkl") 
ori_df = raw_data_preparation(ori_df)

test_reduced_df = pd.read_pickle("./Data/Test_predictions_final(full)_pickl.pkl")

assessment_model = assessment_dataframe_creation(ori_df, test_reduced_df, reference_texts_test_reduced, le)
assessment_model.to_pickle("./Data/assessment_model.pkl")

#pd.set_option('display.max_colwidth',-1)
#assessment_model[assessment_model['model_wer'] < assessment_model['machine_wer']]

## Example

sent = 65
comparison(sent, test_reduced_df, reference_texts_test_reduced, le, assessment_model, ori_df)
