==> 40_analysis.ipynb <==
import pandas as pd
import numpy as np
from tqdm import tqdm
import pickle
import regex as re
import warnings
warnings.filterwarnings("ignore")
import jiwer
from jiwer import wer

from functions import *

# Loading data

model = pd.read_pickle("./Data/assessment_model.pkl") 
voting =  pd.read_pickle("./Data/assessment_voting.pkl") 
voting = voting[voting['reference.text'].isin(model['reference.text'].unique().tolist())]
assessment_df = pd.merge(voting, model, on='reference.text').drop(['machine_wer', 'best_machine_sentence', 'best_machine'], axis=1)
assessment_df['model_class'] = np.where((assessment_df['model_wer'] < assessment_df['min_wer']) , 'better', 
                               np.where((assessment_df['model_wer'] > assessment_df['min_wer']) , 'worse', 'same'))
assessment_df.to_pickle("./Data/assessment_comparison.pkl")

print(assessment_df.shape)
assessment_df.head()

assessment_df['voting_class'].value_counts()

assessment_df['model_class'].value_counts()

model_win = assessment_df[assessment_df['voting_wer'] > assessment_df['model_wer']].shape[0]
voting_win = assessment_df[assessment_df['voting_wer'] < assessment_df['model_wer']].shape[0]
print('Voting better than model: ', voting_win)
print('Model better than voting: ', model_win)

print('WER mean machines: ', assessment_df['min_wer'].mean())
print('WER mean voting  : ', assessment_df['voting_wer'].mean())
print('WER mean model   : ', assessment_df['model_wer'].mean())

# Test prediction Analysis

train = pd.read_pickle("./Data/alignment_with_ref.pkl") 
test = pd.read_pickle("./Data/alignment_test.pkl") 

le, col = word_encoding(train, test)
train[col] = train[col].apply(lambda x :le.transform(x))

test_df = pd.read_pickle("./Data/Test_predictions_final(full)_pickl.pkl")

thresh = 50
test_df_transformed = prediction_data_preparation(test_df, le, thresh)

sentences = test_df['reference.text'].unique().tolist()
predictions_sentences, selected_sentences = sentence_probability(sentences, test_df_transformed)

# example rows where model performs better then best ASR machines
model_better_than_min_wer = assessment_df[assessment_df['model_wer'] < assessment_df['min_wer']].index.tolist()

example_sentence = 60
probability_analysis(sentences, predictions_sentences, selected_sentences, example_sentence, assessment_df)

#test_df_transformed[test_df_transformed['reference.text'] == 'to the fervent latter day saint a temple is not simply a church building a house for religious assembly']

# Creating comparison dataframe

truth = []
best_machine = []
machine_wer = []
model_prediction = []
model_wer = []
thresh = []
thresh_wer = []

for x in tqdm(range(len(sentences))):
    best_machine.append(assessment_df[assessment_df['reference.text'] == sentences[x]]['best_hyp'].tolist()[0][0])    
    truth.append(sentences[x])
    machine_wer.append(wer(sentences[x], best_machine))
    model_prediction.append(predictions_sentences[x])
    model_wer.append(wer(sentences[x], predictions_sentences[x]))
    thresh.append(selected_sentences[x])
    thresh_wer.append(wer(sentences[x], selected_sentences[x]))

compare = pd.DataFrame()
compare['truth'] = truth
compare['machine'] = best_machine
compare['machine_wer'] = machine_wer
compare['model_prediction'] = model_prediction
compare['model_wer'] = model_wer
compare['selected'] = thresh
compare['selected_wer'] = thresh_wer

#compare.to_pickle(path + "/assessment_model(with_thres).pkl")

compare = pd.read_pickle("./Data/assessment_model(with_thres).pkl") 
compare.head()

print('WER mean machines: ', assessment_df['min_wer'].mean())
print('WER mean voting  : ', assessment_df['voting_wer'].mean())
print('WER mean model   : ', assessment_df['model_wer'].mean())
print('WER mean model   : ', compare.selected_wer.mean())
